requirements.txt
matplotlib==3.7
numpy==1.24
Pillow==9.5
pyTelegramBotAPI==4.11
python-dotenv==1.0.0
requests==2.29
schedule==1.2
scipy==1.10
